[FTR][SLOT-107] Add changes from code review

+++ src/main/java/com/three_tech_solutions/slot_app/controllers/responses/StudentDetailsResponse.java
@@ -19,7 +19,7 @@ public record StudentDetailsResponse (
      LocalDate admissionDate,
      String paymentPlanName,
      String plan,
-     Integer numberOfDays,
+     Byte numberOfDays,
      Byte paymentDay,
      boolean status,
      StudentSituation situation,
